% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactPotentialProfiling_FragmentLibrary.R
\name{ContactPotentialProfiling_FragmentLibrary}
\alias{ContactPotentialProfiling_FragmentLibrary}
\alias{CPP_FragmentLibrary}
\title{Generate sequence fragment libraries.}
\usage{
CPP_FragmentLibrary(sequenceSet, windowSizeSet)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}

\item{windowSizeSet}{A set of sliding window sizes.}
}
\description{
Generate sequence fragment libraries.
}
