% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_EpitopeDataset.R
\name{DataPreparation_EpitopeDataset}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Import}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Add_Disease}
\alias{DataPreparation_EpitopeDataset}
\alias{compressedToLongFormat}
\alias{DataPreparation_EpitopeDataset}
\alias{compressedToDummyDF}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Add_HLASerotypes}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Convert_HLASupertypes}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Add_HLARestrictions}
\title{Compile epitope datasets retrieved from IEDB and other sources.}
\usage{
Epitope_Import(IEDBAssayFileName, OtherFileNames = NULL)

Epitope_Add_Disease(df,
  IEDBEpitopeFileName = system.file("IEDB_Epitope_Healthy.csv.gz", package =
  "Repitope"))

compressedToLongFormat(df, compressedColumnName)

compressedToDummyDF(df, compressedColumnName)

Epitope_Add_HLASerotypes(df,
  IEDBEpitopeFileNames = system.file("IEDB_Epitope_Serotype_HLA-A01.csv.gz",
  package = "Repitope"))

Epitope_Convert_HLASupertypes(df, compressedColumnName = "MHC")

Epitope_Add_HLARestrictions(df)
}
\arguments{
\item{IEDBAssayFileName}{A set of T cell assay results obtained from IEDB.}

\item{OtherFileNames}{Other epitope source files. Must contain a "Dataset" column indicating the data source.}

\item{df}{An epitope dataframe.}

\item{IEDBEpitopeFileName}{A set of epitopes obtained from IEDB.}

\item{compressedColumnName}{A string indicating the names of the compressed column to be converted into a long format.}

\item{IEDBEpitopeFileNames}{A set of epitope files from IEDB.}
}
\description{
\code{Epitope_Import} imports IEDB and other epitope files. Contradicting immunogenicity annotations are internally resolved. \cr
\code{Epitope_Add_Disease} adds metadata of being related to diseases or not obtained from the IEDB database.\cr
\code{compressedToLongFormat} converts a compresed column into a long-format column. The compressed strings should be separated by "|".\cr
\code{compressedToDummyDF} converts a compresed column into a dummy variable dataframe. The compressed strings should be separated by "|".\cr
\code{Epitope_Add_HLASerotypes} adds HLA serotype metadata obtained from the IEDB database.\cr
\code{Epitope_Convert_HLASupertypes} converts the compressed HLA genotype information into HLA supertypes. The definition of HLA supertypes is derived from the Additional File 1 of Sidney et al., 2008.\cr
\code{Epitope_Add_HLARestrictions} merges HLA serotypes and HLA supertypes.
}
