% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_NetMHC.R
\name{DataPreparation_NetMHC}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_PeptideExport}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_Import}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_Script}
\title{Utility functions for NetMHC.}
\usage{
NetMHC_PeptideExport(peptideSet, outDir = "./NetMHC/")

NetMHC_Import(netMHCOutputFileNames)

NetMHC_Script(netMHCFastaFileNames, outDir = "./NetMHC/",
  outFileName = "NetMHC_Script.txt")
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{outDir}{A path to the output directory.}

\item{netMHCOutputFileNames}{A set of tab-delimited NetMHC result files.}

\item{netMHCFastaFileNames}{A set of FASTA files for NetMHC analysis.}

\item{outFileName}{A file name for an output.}
}
\description{
Utility functions for NetMHC.
}
