% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_TCR_FastQDump.R
\name{FastQDump_Script}
\alias{FastQDump_Script}
\title{Downloading fastQ files from NCBI Sequence Read Archive (SRA).}
\usage{
FastQDump_Script(accessionSet, outDir = "C:/SRA/output")
}
\arguments{
\item{accessionSet}{A set of SRA Run Accession numbers.}

\item{outDir}{A directory for outputs.}
}
\description{
\code{FastQDump_Script} generates fustq-dump script.
}
