% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_TCR_MiXCR.R
\name{DataPreparation_TCR_MiXCR}
\alias{DataPreparation_TCR_MiXCR}
\alias{MiXCR_Script}
\alias{MiXCR_CombineClonotypes}
\title{MiXCR workflow.}
\usage{
MiXCR_Script(dir = "C:/SRA/output", mixcrPath = "C:/MiXCR/",
  javaMaxMemory = "6G")

MiXCR_CombineClonotypes(dirClones = NULL, clonesFileNames = NULL,
  countSummarise = F, rdsFileName = NULL)
}
\arguments{
\item{dir}{A directory in which FASTQ files are stored.}

\item{mixcrPath}{A path to MiXCR JAR file.}

\item{javaMaxMemory}{A maximum memory size allocated to the Java virtual machine.}

\item{dirClones}{A directory in which MiXCR clonotype TXT files are stored.}

\item{clonesFileNames}{(Optional) Alternativey, MiXCR clonotype TXT filenames can be provided as a list.}

\item{countSummarise}{Logical. If True, the number of datasets containing each CDR3 sequence is counted. If False, raw SRA accessions are provided as such.}

\item{rdsFileName}{(Optional) A filename to which the clonotype summary table is saved as an RDS file.}
}
\description{
\code{MiXCR_Script} generates MiXCR script.\cr
\code{MiXCR_CombineClonotypes} reads and combines MiXCR-derived clonotypes. Note: Only TCR beta chains are retained.\cr
}
