% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_AACPMatrix.R
\docType{data}
\name{Dataset_AACPMatrix}
\alias{Dataset_AACPMatrix}
\alias{AACPMatrix}
\title{Amino acid pairwise contact potential (AACP) matrices.}
\format{A tibble with 35 rows and 2 variables}
\usage{
AACPMatrix
}
\description{
\code{AACPMatrix} is a tibble containing 35 pre-selected AACP matrices.
}
\examples{
dplyr::filter(AACPMatrix, AAIndexID=="AAIndexID-of-your-interest")
}
\keyword{datasets}
