% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R,
%   R/Dataset_FeatureSet.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{MHCI_Human}
\alias{MHCI_Rodents}
\alias{MHCI_Primates}
\alias{MHCII_Human}
\alias{MHCII_Rodents}
\alias{MHCII_Primates}
\alias{MHCI_Human_MinimumFeatureSet}
\alias{MHCII_Human_MinimumFeatureSet}
\title{Compiled datasets of MHC class I and class II epitopes and ligands.}
\format{
data.table

An object of class \code{data.table} (inherits from \code{data.frame}) with 8756 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 411 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 31693 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 8445 rows and 9 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 9 columns.

Character vectors containing 32 and 27 feature names for HLA-I and HLA-II, respectively.

An object of class \code{character} of length 27.
}
\usage{
MHCI_Human

MHCI_Rodents

MHCI_Primates

MHCII_Human

MHCII_Rodents

MHCII_Primates

MHCI_Human_MinimumFeatureSet

MHCII_Human_MinimumFeatureSet
}
\description{
Datasets were compiled from source files attached in the package using \code{Epitope_Import}.
}
\keyword{datasets}
