% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset_Mouse.R
\docType{data}
\name{Dataset_EpitopeDataset_Mouse}
\alias{Dataset_EpitopeDataset_Mouse}
\alias{EpitopeDataset_Mouse}
\title{A compiled dataset of mouse MHC class I epitopes and ligands.}
\format{A tibble with 17558 rows and 7 variables}
\usage{
EpitopeDataset_Mouse
}
\description{
A compiled dataset of mouse MHC class I epitopes and ligands.
}
\keyword{datasets}
