% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset_Primates.R
\docType{data}
\name{Dataset_EpitopeDataset_Primates}
\alias{Dataset_EpitopeDataset_Primates}
\alias{EpitopeDataset_Primates}
\title{A compiled dataset of primates MHC class I epitopes and ligands.}
\format{A tibble with 536 rows and 7 variables}
\usage{
EpitopeDataset_Primates
}
\description{
A compiled dataset of primates MHC class I epitopes and ligands.
}
\keyword{datasets}
