% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_DecodeParameters.R
\name{Features_DecodeParameters}
\alias{Features_DecodeParameters}
\title{Decode parameters used for calculating the minimum set of features.}
\usage{
Features_DecodeParameters(preprocessedDFList_reference)
}
\arguments{
\item{preprocessedDFList_reference}{A list of feature dataframes generated by \code{Features_Preprocess} for reference.}
}
\description{
\code{Features_DecodeParameters} obtains a set of parameters required for calculation of the minimum set of features used in \code{preprocessedDFList_reference}.\cr
\code{Features_Minimum} allows users to generate a set of preprocessed feature dataframes for different sets of peptides and TCR repertoires based on the parameters obtianed by \code{Features_DecodeParameters}.
}
