% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_ErrorCorrect.R
\name{Features_ErrorCorrect}
\alias{Features_ErrorCorrect}
\title{Correct errors in the peptide feature dataframes.}
\usage{
Features_ErrorCorrect(featureDFList, TCRSet, aaIndexIDSet = "all",
  alignTypeSet = "global-local", fragLenSet = 4:6,
  TCRFragDepthSet = 10000, seedSet = 1:5)
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a list, it must be the same length with the seedSet.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched.}

\item{seedSet}{A set of random seeds.}
}
\description{
Correct errors in the peptide feature dataframes.
}
