% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_ExtractRandomComponent.R
\name{Features_ExtractRandomComponent}
\alias{Features_ExtractRandomComponent}
\title{Extract features generated from random sequences}
\usage{
Features_ExtractRandomComponent(featureDFList)
}
\arguments{
\item{featureDFList}{A set of feature dataframes generated by the \code{Features} function.}
}
\description{
Extract features generated from random sequences
}
