% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_IO.R
\name{Features_IO}
\alias{Features_IO}
\alias{readFeatureDFList}
\alias{Features_IO}
\alias{saveFeatureDFList}
\title{Feature dataframes In/Out.}
\usage{
readFeatureDFList(featureDFFileNameList)

saveFeatureDFList(featureDFList, fileNameHeader)
}
\arguments{
\item{featureDFFileNameList}{A set of binary file names with ".fst" extensions.}

\item{featureDFList}{A named list of feature dataframes.}

\item{fileNameHeader}{A file name header.}
}
\description{
Feature dataframes In/Out.
}
