% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Minimum.R
\name{Features_Minimum}
\alias{Features_Minimum}
\title{Calculate the minimum set of features.}
\usage{
Features_Minimum(peptideSet, TCRSet, metadataDFList,
  preprocessedDFList_reference, coreN = parallel::detectCores())
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a list, it must be the same length with the seedSet.}

\item{metadataDFList}{A list of metadata dataframes labeled as "Peptide", "Immunogenicity", and "Cluster". No extra metadata is allowed. Feature dataframes are downsized based on the "Peptide" column in this metadata. In other words, peptides not contained in the metadata dataframe are omitted.}

\item{preprocessedDFList_reference}{A list of feature dataframes generated by \code{Features_Preprocess} for reference.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
\code{Features_DecodeParameters} obtains a set of parameters required for calculation of the minimum set of features used in \code{preprocessedDFList_reference}.\cr
\code{Features_Minimum} allows users to generate a set of preprocessed feature dataframes for different sets of peptides and TCR repertoires based on the parameters obtianed by \code{Features_DecodeParameters}.
}
