% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_MinimumParameters.R
\name{Features_MinimumParameters}
\alias{Features_MinimumParameters}
\title{Decode parameters used for calculating the minimum set of features.}
\usage{
Features_MinimumParameters(preprocessedDFList, criteria = "intersect")
}
\arguments{
\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess}, \code{Features_CorFilter}, or \code{Features_FeatureSelect} for reference.}

\item{criteria}{Can either be "intersect" or "union".}
}
\description{
\code{Features_MinimumParameters} obtains a set of parameters required for calculation of the minimum set of features used in \code{preprocessedDFList}.\cr
}
