% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Mock.R
\name{Features_Mock}
\alias{Features_Mock}
\title{Extract information generated from randomly generated mock sequences}
\usage{
Features_Mock(featureDFList)
}
\arguments{
\item{featureDFList}{A set of feature dataframes generated by the \code{Features} function.}
}
\description{
Extract information generated from randomly generated mock sequences
}
