% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_PeptideDescriptor.R
\name{Features_PeptideDescriptor}
\alias{Features_PeptideDescriptor}
\title{Descriptive statistics of peptide fragment descriptors.}
\usage{
Features_PeptideDescriptor(peptideSet, fragLenSet = 3:8)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}
}
\description{
\code{Features_PeptideDescriptor} calculates descriptive statistics using peptide descriptors.\cr
\code{Features_rTPCP} calculates descriptive statistics of repertoire-wide TCR-peptide pairwise contact potentials.\cr
\code{Features} is a wrapper function, generating a list of dataframes of peptide features calculated with different parameter sets.
}
