% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocess.R
\name{Features_Preprocess}
\alias{Features_Preprocess}
\title{Preprocess peptide feature dataframes for machine learning.}
\usage{
Features_Preprocess(featureDFList, metadataDF)
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}. Can also be a merged output from \code{Features_Merge_} functions.}

\item{metadataDF}{A metadata dataframe labeled as "Peptide", "Immunogenicity", and "Cluster". No extra metadata is allowed. Feature dataframes are downsized based on the "Peptide" column in this metadata. In other words, peptides not contained in the metadata dataframe are omitted.}
}
\description{
Preprocess peptide feature dataframes for machine learning.
}
