% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocessing.R
\name{Features_Preprocessing}
\alias{Features_Preprocessing}
\alias{Features_Preprocess}
\alias{Features_Preprocessing}
\alias{Features_CorFilter}
\alias{Features_Preprocessing}
\alias{Features_FeatureSelect}
\title{Preprocess peptide feature dataframes for machine learning.}
\usage{
Features_Preprocess(featureDFList, metadataDF, seedSet = 1:5)

Features_CorFilter(preprocessedDFList, corThreshold = 0.75,
  coreN = parallel::detectCores())

Features_FeatureSelect(preprocessedDFList, featureN = 100,
  coreN = parallel::detectCores())
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{metadataDF}{A dataframe containing metadata, consisting of "Peptide", "Immunogenicity", and "Cluster" columns.}

\item{seedSet}{A set of random seeds.}

\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess} or \code{Features_CorFilter}.}

\item{corThreshold}{The threshold of correlation to eliminate features.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to disable.}

\item{featureN}{The number of features to be retained.}
}
\description{
Preprocess peptide feature dataframes for machine learning.
}
