% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneTransition.R
\name{ImmuneTransition}
\alias{ImmuneTransition}
\alias{ImmuneTransition_TrainModels}
\alias{ImmuneTransition_SummarizeInternalScores}
\alias{ImmuneTransition_Score}
\alias{ImmuneTransition_Predict}
\title{Immune transition score.}
\usage{
ImmuneTransition_TrainModels(featureDF, metadataDF, featureSet = "all",
  seedSet = 1:5, coreN = parallel::detectCores(logical = F))

ImmuneTransition_SummarizeInternalScores(trainModelResults)

ImmuneTransition_Score(featureDF, metadataDF, featureSet = "all",
  seedSet = 1:5, coreN = parallel::detectCores(logical = F))

ImmuneTransition_Predict(externalFeatureDFList, trainModelResults)
}
\arguments{
\item{featureDF}{A dataframe of features generated by \code{Features}.}

\item{metadataDF}{A dataframe containing metadata, consisting of "Peptide" and "ImmuneTransition" columns. Other columns if provided would be used as features.}

\item{featureSet}{A set of features used for model training. Set "all" to shortcut.}

\item{seedSet}{A set of random seeds.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to disable.}

\item{trainModelResults}{The model training result returned by \code{ImmuneTransition_TrainModels}.}

\item{externalFeatureDFList}{The feature dataframes for which immune transition is to be predicted.}
}
\description{
\code{ImmuneTransition_TrainModels} performes preprocessing, trains extremely randomized tree models, and internally predicts immune transition scores on the entire dataset provided.\cr
\code{ImmuneTransition_SummarizeInternalScores} summarizes internally calculated immune transition scores.\cr
\code{ImmuneTransition_Score} is a wrapper function of \code{ImmuneTransition_TrainModels} and \code{ImmuneTransition_SummarizeInternalScores}.\cr
\code{ImmuneTransition_Predict} predicts immune transition scores on external datasets. Models trained by \code{ImmuneTransition_TrainModels} are necessary. Note that, due to Java constraints, models have to be constructed in each new R session.
}
