% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity.R
\name{Immunogenicity}
\alias{Immunogenicity}
\title{Immunogenicity prediction.}
\usage{
Immunogenicity(preprocessedDFList = NULL, featureSet = "all",
  destDir = "./Immunogenicity/", outputHeader = "Output",
  coreN = parallel::detectCores())
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess}, \code{Features_CorFilter}, or \code{Features_FeatureSelect}.}

\item{featureSet}{A set of features to be used for model training. If "all", all features in the training dataframe except "DataType", "Immunogenicity", "Peptide", and "Cluster" will be used.}

\item{destDir}{A directory for loading & exporting data.}

\item{outputHeader}{A file/folder name header for the outputs to be saved.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
\code{Immunogenicity} is a wrapper function for training and evaluating a classifier. The optimized algorithm is used by default.\cr
}
