% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity_Benchmark.R
\name{Immunogenicity_Benchmark}
\alias{Immunogenicity_Benchmark}
\title{Benchmarking various algorithms for immunogenicity prediction.}
\usage{
Immunogenicity_Benchmark(preprocessedDFList, learnerSet = NULL,
  omitLearnerSet = NULL, destDir = "./Benchmark/",
  coreN = parallel::detectCores())
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess}, \code{Features_CorFilter}, or \code{Features_FeatureSelect}.}

\item{learnerSet}{(Optional) A set of learner strings to be used for model benchmarking. Set \code{NULL} to ignore.}

\item{omitLearnerSet}{(Optional) A set of learner strings to be omitted from model benchmarking. Set \code{NULL} to ignore.}

\item{destDir}{A directory for loading & exporting data.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}

\item{featureSet}{A set of features to be used for model training. If "all", all features in the training dataframe except "DataType", "Immunogenicity", "Peptide", and "Cluster" will be used.}

\item{outputHeader}{A file/folder name header for the outputs to be saved.}
}
\description{
\code{Immunogenicity_Benchmark} is a wrapper function for benchmarking classifiers implemented in the \code{mlr} package.\cr
}
