% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity_Score.R
\name{Immunogenicity_Score}
\alias{Immunogenicity_Score}
\title{Immunogenicity score.}
\usage{
Immunogenicity_Score(featureDFFileNames_train,
  featureDFFileNames_predict = NULL, metadata_train, featureSet = "all",
  seedSet = 1:10, coreN = parallel::detectCores())
}
\arguments{
\item{featureDFFileNames_train}{A character vector of the fst file names of feature dataframes for model training.}

\item{featureDFFileNames_predict}{A character vector of the fst file names of feature dataframes for predicting immunogenicity scores. If the length equals to that of \code{featureDFFileNames_train}, each dataframe will be subject to the prediction by the classifier trained from the corresponding training dataframe. Otherwise, they will be subject to the prediction in an all-by-all manner. If \code{NULL}, model training dataframe will be splitted into training and prediction dataframes.}

\item{metadata_train}{A dataframe containig "Peptide" and "Immunogenicity" columns for model training.}

\item{featureSet}{A set of features for model training.}

\item{seedSet}{A set of random seeds for bootstrapping. Can be a single value when \code{featureDFFileNames_predict} is provided.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
Immunogenicity score.
}
