% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmunogenicityPrediction.R
\name{ImmunogenicityPrediction}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity_AutoML}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity_Prediction}
\title{Immunogenicity prediction using H2O AutoML instances}
\usage{
Immunogenicity(preprocessedDFList, featureSet = "all",
  destDir = "./Results/", outputHeader = "Output", max_mem_size = "6G",
  nthreads = 6)

Immunogenicity_AutoML(trainDF, testDF, validDF, featureSet = "all",
  destDir = "./Results/", LeaderBoardName = "LeaderBoard.csv",
  H2OModelName = "BestH2OModel", seed = 12345, max_mem_size = "6G",
  nthreads = 6)

Immunogenicity_Prediction(evalDF, destDir = "./Results/",
  H2OModelName = "BestH2OModel", PredictionHeader = "Predictions",
  seed = 12345, max_mem_size = "6G", nthreads = 6)
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess} or \code{Features_FeatureSelection}. A nested list for batch processing is allowed.}

\item{featureSet}{A set of features to be used for model training. If "all", all features in the training dataframe except "DataType", "Immunogenicity", "Peptide", and "Cluster" will be used.}

\item{destDir}{A directory either from importing H2O models or for exporting outputs.}

\item{outputHeader}{A file/folder name header.}

\item{max_mem_size}{The upper limit of memory for H2O Java machine learning engine.}

\item{nthreads}{The number of threads for H2O Java machine learning engine.}

\item{trainDF}{A dataframe for model training.}

\item{testDF}{A dataframe for hold-out validation.}

\item{validDF}{A dataframe for leaderboard scoring.}

\item{LeaderBoardName}{The file name of the leader board to be exported. The file format must be CSV. If you want to skip exporting, set \code{NULL}.}

\item{H2OModelName}{The file name of the best H2O model to be exported. No file extension is required. If you want to skip exporting, set \code{NULL}.}

\item{seed}{A random seed.}

\item{evalDF}{A dataframe for evaluation.}

\item{PredictionHeader}{The file name header of the prediction results to be exported.}
}
\description{
\code{Immunogenicity} is a wrapper function.\cr
\code{Immunogenicity_AutoML} conducts automated machine learning and exports the best H2O binary model.\cr
\code{Immunogenicity_Prediction} makes predictions using the previously exported model.
}
