% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningCurve.R
\name{learningCurve}
\alias{learningCurve}
\title{Learning curve analysis for the caret machine learning pipeline}
\usage{
learningCurve(trainingData.preProc, testingData.preProc, outcomeLabelName,
  mlAlgorithm = "xgbLinear", trainControlOptions, weighted = F,
  proportionSet = (1:10)/10, seed = 12345)
}
\arguments{
\item{trainingData.preProc}{A pre-processed dataframe for model training.}

\item{testingData.preProc}{A pre-processed dataframe for model testing.}

\item{outcomeLabelName}{A string for the outcome label name.}

\item{mlAlgorithm}{A string indicating the caret machine learning algorithm.}

\item{trainControlOptions}{A trainControl object.}

\item{weighted}{Logical. Whether class imbalance should be adjusted by class-specific weighting.}

\item{proportionSet}{A numeric vector indicating the set of proportions for learning curve analysis.}

\item{seed}{A random seed.}

\item{...}{Arguments passed to \code{caret::train}}
}
\description{
Learning curve analysis for the caret machine learning pipeline
}
