% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcBinderPrediction.R
\name{mhcBinderPrediction}
\alias{mhcBinderPrediction}
\title{Prediction of MHC binders using the stabilized matrix method (SMM) approach}
\usage{
mhcBinderPrediction(peptideSet, ic50.threshold = 500)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{ic50.threshold}{Peptides with predicted IC50 values lower than this will be considered binders. A threshold of 500 nM has been used frequently, according to Sette et al. https://www.ncbi.nlm.nih.gov/pubmed/7527444}
}
\description{
returns peptides that bind to at least one of the 50 HLA alleles implemented in the package \code{EpitopePrediction}.
}
