% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXCR.R
\name{MiXCR}
\alias{MiXCR}
\alias{MiXCRScript}
\alias{MiXCR}
\alias{MiXCRClonotypeCombine}
\title{MiXCR workflow.}
\usage{
MiXCRScript(dir = "C:/SRA/output", mixcrPath = "C:/MiXCR/",
  javaMaxMemory = "6G")

MiXCRClonotypeCombine(dir, outDir, rdsFileName = "TCR_Combined.rds",
  countSummarise = F)
}
\arguments{
\item{dir}{A directory in which FASTQ files are stored.}

\item{mixcrPath}{A path to MiXCR JAR file.}

\item{javaMaxMemory}{A maximum memory size allocated to the Java virtual machine.}

\item{outDir}{A directory in which the clonotype summary table is saved.}

\item{rdsFileName}{A filename for the clonotype summary.}

\item{countSummarise}{A logical. If True, the number of datasets containing each CDR3 sequence is counted. If False, raw SRA accessions are provided as such.}
}
\description{
\code{MiXCRScript} generates MiXCR script.\cr
\code{MiXCRClonotypeCombine} reads and combines MiXCR-derived clonotypes. Note: Only TCR beta chains are retained.
}
