% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{multiPlot}
\alias{multiPlot}
\title{Combine multiple ggplots}
\usage{
multiPlot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplots}

\item{plotlist}{A list of ggplots}

\item{cols}{The number of columns}

\item{layout}{A matrix for layouting the plots}
}
\description{
\code{multiplot} combines multiple ggplots.
}
