% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborNetwork.R
\name{NeighborNetwork}
\alias{NeighborNetwork}
\alias{distMat_Sbst}
\alias{NeighborNetwork}
\alias{distMat_Indel}
\alias{NeighborNetwork}
\alias{neighborNetwork}
\title{Neighbor network analysis.}
\usage{
distMat_Sbst(peptideSet)

distMat_Indel(longerPeptideSet, shorterPeptideSet)

neighborNetwork(peptideSet, numSet = NULL, directed = T, weighted = T,
  forceMutType = T)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{longerPeptideSet}{A set of peptide sequences. Should be one amino acid longer than \code{shorterPeptideSet}.}

\item{shorterPeptideSet}{A set of peptide sequences. Should be one amino acid shorter than \code{longerPeptideSet}.}

\item{numSet}{An attribute for the vertices.}

\item{directed}{Should the network be converted from undirected to directed? Directions are determined by the \code{numSet} provided.}

\item{weighted}{Should the network be converted to weihted? Edge weights are determined by the \code{numSet} provided.}

\item{forceMutType}{Should mutational types be annotated? A little bit time-consuming.}
}
\description{
Neighbor network analysis.
}
