% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborNetwork_ImmuneEscapePotential.R
\name{NeighborNetwork_ImmuneEscapePotential}
\alias{NeighborNetwork_ImmuneEscapePotential}
\alias{neighborNetwork_ClusterPlot}
\title{Immune escape potentials derived from neighbor network clustering analysis.}
\usage{
neighborNetwork_ClusterPlot(peptide, neighborNetResult, metadataDF,
  seed = 12345)
}
\arguments{
\item{peptide}{The target peptide sequence.}

\item{neighborNetResult}{The result returned from \code{neighborNetwork}.}

\item{metadataDF}{A dataframe which has "Peptide", "Immunogenicity", and "ImmunogenicityScore" columns.}

\item{seed}{A random seed.}
}
\description{
Immune escape potentials derived from neighbor network clustering analysis.
}
