% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborNetwork_InsillicoMutagenesis.R
\name{NeighborNetwork_InsillicoMutagenesis}
\alias{NeighborNetwork_InsillicoMutagenesis}
\alias{InSillicoMutagenesis}
\alias{NeighborNetwork_InsillicoMutagenesis}
\alias{InsillicoMutagenesis_GenerateIntermediates}
\alias{NeighborNetwork_InsillicoMutagenesis}
\alias{InsillicoMutagenesis_PairwiseNeighborNetwork}
\title{In sillico mutagenesis analysis.}
\usage{
InSillicoMutagenesis(peptideSet, coreN = NULL)

InsillicoMutagenesis_GenerateIntermediates(peptidePair)

InsillicoMutagenesis_PairwiseNeighborNetwork(peptidePairDF,
  coreN = parallel::detectCores())
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{coreN}{The number of threads for parallelization.}

\item{peptidePair}{A pair of peptide sequences.}

\item{peptidePairDF}{A dataframe which has two columns of paired peptide sequences.}
}
\description{
\code{InSillicoMutagenesis} generates single-mutated peptide sequences.
\code{InsillicoMutagenesis_GenerateIntermediates} generates mutational intermediates between two peptide sequences.
\code{InsillicoMutagenesis_PairwiseNeighborNetwork} generates single-mutated peptide sequences and construct neighbor networks.
}
