% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neoepitope.R
\name{Neoepitope}
\alias{Neoepitope}
\alias{Neoepitope_BurdenDT}
\alias{Neoepitope}
\alias{Neoepitope_PValueDT}
\alias{Neoepitope}
\alias{Neoepitope_PValueDT_Batch}
\alias{Neoepitope}
\alias{Neoepitope_PValueLineChart}
\alias{Neoepitope}
\alias{Neoepitope_KMPlot}
\title{Neoepitope analysis.}
\usage{
Neoepitope_BurdenDT(dt_neoepitope, thr = "none", thr.ne = 1)

Neoepitope_PValueDT(dt_neoepitope, thr = "none")

Neoepitope_PValueDT_Batch(dt_neoepitope, thrSet = seq(0, 1, length.out =
  1000), coreN = parallel::detectCores())

Neoepitope_PValueLineChart(dt_pvalue)

Neoepitope_KMPlot(dt_neoepitope, thr = "none", thr.ne = 1)
}
\arguments{
\item{dt_neoepitope}{A datatable of neoepitopes that contains the following columns: "Dataset","Sample","Months","Status", "ImmunogenicityScore.WT", and, "ImmunogenicityScore.MT".}

\item{thr}{A threshold of neoepitope dissimilarity index. Can be set as "none" to disabled. In that case, neoepitope dissimilarity indices were aggregated by patients without thresholding.}

\item{thr.ne}{A threshold of neoepitope burden or aggregated neoepitope dissimilarity index per patient.}

\item{thrSet}{A set of thresholds of neoepitope dissimilarity index.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to disable parallelization.}

\item{dt_pvalue}{A datatable containing thresholds and P-values by the log-rank test.}
}
\description{
\code{Neoepitope_BurdenDT} returns a neoepitope summary datatable using the thresholds provided. \cr
\code{Neoepitope_PValueDT} calculates P-values by the log-rank test with varying thresholds for the neoepitope burden defined by the specified threshold of the neoepitope dissimilarity index. If no threshold is provided, neoepitope dissimilarity indices were aggregated by patients without thresholding. \cr
\code{Neoepitope_PValueDT_Batch} does the jobs for a set of thresholds of the neoepitope dissimilarity index. \cr
\code{Neoepitope_PValueLineChart} generates a line chart of threshold values and P-values.\cr
\code{Neoepitope_KMPlot} generates a Kaplan-Mayer curve with the thresholds provided.
}
