% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_PreTrained}
\alias{predict_PreTrained}
\title{Predict immunogenicity using pre-trained classifiers}
\usage{
predict_PreTrained(variableDataFrame, variableType)
}
\arguments{
\item{variableDataFrame}{A dataframe containing either rTPCP or mrTPCP variables. Currently, only MIYS990106-derived variables with window sizes of 4 and 5 are supported. In case of mrTPCP, currently only nonapeptides are supported.}

\item{variableType}{Either "rTPCP" or "mrTPCP".}
}
\description{
Predict immunogenicity from the rTPCP/mrTPCP matrix calculated from user-provided peptide sequences using pre-trained classifiers
}
