% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicTCRAnalysis.R
\name{publicTCRAnalysis}
\alias{publicTCRAnalysis}
\alias{similarityNetworkCentralNodes}
\alias{publicTCRAnalysis}
\alias{similarityNetworkCentralNodes.Large}
\alias{publicTCRAnalysis}
\title{Public TCR clonotypes contributing as central nodes in a similarity network}
\usage{
similarityNetworkCentralNodes(TCRSet)

similarityNetworkCentralNodes.Large(largeTCRSet, n.max.clonotype = 10000,
  n.max.iteration = 20, seed = 12345)

publicTCRAnalysis(combinedTCRDF, n.data = 2, n.max.clonotype = 10000,
  n.max.iteration = 20, seed = 12345, outDir, header = "TCR_")
}
\arguments{
\item{TCRSet}{A set of TCR sequences.}

\item{largeTCRSet}{A large set of TCR sequences.}

\item{n.max.clonotype}{An integer representing the maximum number of clonotypes for construction of similarity network.}

\item{n.max.iteration}{An integer representing the maximum number of iterations.}

\item{seed}{A random seed.}

\item{combinedTCRDF}{A dataframe of TCR clonotypes. Must contain columns labeled as "aaSeqCDR3" and "N". Column "N" represents the number of datasets in which the CDR3 sequence were observed.}

\item{n.data}{An integer representing the threshold of publicity. TCRs observed in at least \code{n.data} repertoire datasets are considered public.}

\item{outDir}{A directory in which the public clonotype analysis results are saved.}

\item{header}{A filename header for public clonotype analysis results.}
}
\description{
\code{similarityNetworkCentralNodes} is a working function to extract central nodes from TCR CDR3 similarity network, where any pair of clonotypes with only one amino acid difference (levenstein distance of one) is considered linked.\cr
\code{similarityNetworkCentralNodes.Large} randomely splits the input TCR set, performs central node analysis, extracts important TCRs in each similarity networks, and finally performs central node analysis again with the set of important TCRs.\cr
\code{publicTCRAnalysis} is a wrapper function.
}
