% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_csv_checkColumnNames}
\alias{read_csv_checkColumnNames}
\title{Reading a CSV file and remove spaces from the column names}
\usage{
read_csv_checkColumnNames(fileName)
}
\arguments{
\item{fileName}{a CSV file name}
}
\description{
Reading a CSV file and remove spaces from the column names
}
