% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveCurrentGraphicPDF}
\alias{saveCurrentGraphicPDF}
\title{Plot saving utilities}
\usage{
saveCurrentGraphicPDF(outputFileName, width, height, pointsize = 12,
  useDingbats = F, ghostScriptPath = "C:/gs/gs9.16/bin/gswin32c.exe")
}
\arguments{
\item{outputFileName}{The name of the exported PDF file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{pointsize}{A pointsize}

\item{useDingbats}{Logical indicating whether points shuold be replaced with the Dingbat font.}

\item{ghostScriptPath}{The file path to the GhostScript.exe file.}
}
\description{
\code{saveCurrentGraphicPDF} saves the plot currently on the graphics device to a PDF file.
}
