% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{savePDF}
\alias{savePDF}
\title{Plot saving utilities}
\usage{
savePDF(graphicObject = ggplot2::last_plot(), outputFileName, width, height,
  pointsize = 12, useDingbats = F,
  ghostScriptPath = "C:/gs/gs9.16/bin/gswin32c.exe")
}
\arguments{
\item{graphicObject}{The plot object to be saved.}

\item{outputFileName}{The name of the exported PDF file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{pointsize}{A pointsize}

\item{useDingbats}{Logical indicating whether points shuold be replaced with the Dingbat font.}

\item{ghostScriptPath}{The file path to the GhostScript.exe file.}
}
\description{
\code{savePDF} saves a plot to a PDF file.
}
