% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Publication-ready ggplot scales}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{No extra arguments are supported.}
}
\description{
Publication-ready ggplot scales.
}
