% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-io.R
\name{sequenceFilter}
\alias{sequenceFilter}
\title{Filter amino acid sequences to exclude those containing non-standard letters.}
\usage{
sequenceFilter(sequenceSet)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}
}
\description{
Filter amino acid sequences to exclude those containing non-standard letters.
}
