% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRAnalysis_CentralClonotypes.R
\name{TCRAnalysis_CentralClonotypes}
\alias{TCRAnalysis_CentralClonotypes}
\alias{centralClonotypeAnalysis}
\title{"Central" clonotypes in the similarity repertoire network}
\usage{
centralClonotypeAnalysis(simNet, sizeSet = 10000,
  outputFileHeader = "./Graph/Graph_")
}
\arguments{
\item{simNet}{A similarity network object.}

\item{sizeSet}{A set of integers specifying the minimum numbers of clonotypes retained as central clonotypes. Based on the degree distribution, appropriate degree threshold will be internally calculated.}

\item{outputFileHeader}{A path for saving the outputs from the central clonotype analysis.}
}
\description{
"Central" clonotypes in the similarity repertoire network
}
