% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRAnalysis_FastQDump.R
\name{fastqDumpScript}
\alias{fastqDumpScript}
\title{Downloading fastQ files from NCBI Sequence Read Archive (SRA).}
\usage{
fastqDumpScript(accessionSet, outDir = "C:/SRA/output")
}
\arguments{
\item{accessionSet}{A set of SRA Run Accession numbers.}

\item{outDir}{A directory for outputs.}
}
\description{
\code{fastqDumpScript} generates fustq-dump script.
}
