% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRAnalysis_MiXCR.R
\name{TCRAnalysis_MiXCR}
\alias{TCRAnalysis_MiXCR}
\alias{MiXCRScript}
\alias{TCRAnalysis_MiXCR}
\alias{MiXCRClonotypeCombine}
\title{MiXCR workflow.}
\usage{
MiXCRScript(dir = "C:/SRA/output", mixcrPath = "C:/MiXCR/",
  javaMaxMemory = "6G")

MiXCRClonotypeCombine(dirClones = NULL, clonesFileNames = NULL,
  countSummarise = F, rdsFileName = NULL)
}
\arguments{
\item{dir}{A directory in which FASTQ files are stored.}

\item{mixcrPath}{A path to MiXCR JAR file.}

\item{javaMaxMemory}{A maximum memory size allocated to the Java virtual machine.}

\item{dirClones}{A directory in which MiXCR clonotype TXT files are stored.}

\item{clonesFileNames}{(Optional) Alternativey, MiXCR clonotype TXT filenames can be provided as a list.}

\item{countSummarise}{Logical. If True, the number of datasets containing each CDR3 sequence is counted. If False, raw SRA accessions are provided as such.}

\item{rdsFileName}{(Optional) A filename to which the clonotype summary table is saved as an RDS file.}
}
\description{
\code{MiXCRScript} generates MiXCR script.\cr
\code{MiXCRClonotypeCombine} reads and combines MiXCR-derived clonotypes. Note: Only TCR beta chains are retained.\cr
}
