% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_Publication}
\alias{theme_Publication}
\title{The publication-ready ggplot theme}
\usage{
theme_Publication(base_size = 18, base_family = "Helvetica")
}
\arguments{
\item{base_size}{The base font size.}

\item{base_family}{The base font family.}
}
\description{
The publication-ready ggplot theme.
}
