% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Utility}
\alias{Utility}
\alias{sequenceFilter}
\alias{Utility}
\alias{compressedToLongFormat}
\alias{Utility}
\alias{standardizeHLAString}
\title{Miscellaneous utility functions.}
\usage{
sequenceFilter(sequenceSet)

compressedToLongFormat(df, compressedColumnName)

standardizeHLAString(HLAStrings)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}

\item{df}{A dataframe like \code{Repitope::EpitopeDataset} which has compressed columns.}

\item{compressedColumnName}{A string indicating the names of the compressed column to be converted into a long format.}

\item{HLAStrings}{A character vector of HLA class I strings to be standardized.}
}
\description{
\code{sequenceFilter} filters amino acid sequences so that those containing non-standard letters are excluded.
\code{compressedToLongFormat} converts a compresed dataframe into a long-format dataframe. The compressed strings should be separated by "|".
\code{standardizeHLAString} standardizes HLA class I strings.
}
