% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_DF.R
\name{Utility_DF}
\alias{Utility_DF}
\alias{sortByMedian}
\title{Utility functions for dataframe/datatables.}
\usage{
sortByMedian(df, groupColumnName = "FeatureID",
  valueColumnName = "Importance", descending = F)
}
\arguments{
\item{df}{Any dataframe.}

\item{groupColumnName}{The column name for the grouping variables.}

\item{valueColumnName}{The column name for values.}

\item{descending}{Logical. Whether the dataframe should be reordered in a descending order.}
}
\description{
Utility functions for dataframe/datatables.
}
