% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_FastUnique.R
\name{Utility_FastUnique}
\alias{Utility_FastUnique}
\alias{fastUnique}
\title{The fastest "unique" function.}
\usage{
fastUnique(x)
}
\arguments{
\item{x}{Any vector.}
}
\description{
The fastest "unique" function.
}
