% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParCor.R
\name{Utility_ParCor}
\alias{Utility_ParCor}
\alias{parCor}
\title{Parallelized correlation matrix calculation.}
\usage{
parCor(data, center = T, scale = T, num_splits = parallel::detectCores(),
  verbose = 2)
}
\arguments{
\item{data}{A matrix, a data.frame, or a data.table.}

\item{center}{A logical. Passed to \code{base::scale} for centering the input data.}

\item{scale}{A logical. Passed to \code{base::scale} for scaling the input data.}

\item{num_splits}{The number of chunk blocks. Default setting is the number of core (for parallelization).}

\item{verbose}{Either 0, 1, 2, or TRUE/FALSE.}
}
\description{
Parallelized correlation matrix calculation.
}
