% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot}
\alias{Utility_Plot}
\alias{savePDF}
\title{Plot saving utilities}
\usage{
savePDF(graphicObject = NULL, outputFileName = "plot.pdf", width = 8,
  height = 5, pointsize = 12, useDingbats = F)
}
\arguments{
\item{graphicObject}{The plot object to be saved. If missing, the most recently depicted plot will be saved.}

\item{outputFileName}{The name of the exported PDF file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{pointsize}{A pointsize}

\item{useDingbats}{Logical indicating whether points shuold be replaced with the Dingbat font.}
}
\description{
Plot saving utilities
}
