% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{Utility_Plot_ggplot-scales}
\alias{Utility_Plot_ggplot-scales}
\alias{scale_color_Publication}
\alias{Utility_Plot_ggplot-scales}
\alias{scale_colour_Publication}
\alias{Utility_Plot_ggplot-scales}
\alias{scale_fill_Publication}
\title{Publication-ready ggplot scales}
\usage{
scale_color_Publication()

scale_colour_Publication()

scale_fill_Publication()
}
\description{
Publication-ready ggplot scales
}
