% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Plot.R
\name{theme_Publication}
\alias{theme_Publication}
\title{The publication-ready ggplot theme}
\usage{
theme_Publication(base_size = 18)
}
\arguments{
\item{base_size}{The base font size.}
}
\description{
The publication-ready ggplot theme
}
