% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_SimilarityNetwork.R
\name{Utility_SimilarityNetwork}
\alias{Utility_SimilarityNetwork}
\alias{distMat_Auto}
\alias{Utility_SimilarityNetwork}
\alias{distMat_Juxtaposed}
\alias{Utility_SimilarityNetwork}
\alias{singleAASimilarityNetwork}
\title{Similarity network analysis.}
\usage{
distMat_Auto(aaStringSet)

distMat_Juxtaposed(longerAAStringSet, shorterAAStringSet)

singleAASimilarityNetwork(aaStringSet, numSet = NULL, directed = T,
  weighted = T, forceMutType = F)
}
\arguments{
\item{aaStringSet}{A set of XCR clonotype sequences (converted as an AAStringSet object).}

\item{longerAAStringSet}{A set of XCR clonotype sequences. Should be one amino acid longer than \code{shorterAAStringSet}.}

\item{shorterAAStringSet}{A set of XCR clonotype sequences. Should be one amino acid shorter than \code{longerAAStringSet}.}

\item{numSet}{An attribute for the verteces.}

\item{directed}{Should the network be converted from undirected to directed? Directions are determined by the \code{numSet} provided.}

\item{weighted}{Should the network be converted to weihted? Edge weights are determined by the \code{numSet} provided.}

\item{forceMutType}{Should mutational types be annotated? A little bit time-consuming.}
}
\description{
Similarity network analysis.
}
