% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_VennDiagram.R
\name{Utility_VennDiagram}
\alias{Utility_VennDiagram}
\alias{vennDiagram}
\title{A utility function for creating a Venn diagram.}
\usage{
vennDiagram(x, colors = (ggsci::pal_d3())(length(x)),
  show_category_names = T)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding to a separate circle in the Venn diagram. The length should fall into the range between 1 and 5.}

\item{colors}{A set of colors for the venn diagram.}

\item{show_category_names}{Logical. Whether the names of \code{x} should be printed?}
}
\description{
A utility function for creating a Venn diagram.
}
