% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AACP.R
\name{AACP.AAIndex.Matrix.Format}
\alias{AACP.AAIndex.Matrix.Format}
\title{Format AAIndex Amino Acid Pairwise Contact Potential (AACP) matrix}
\usage{
AACP.AAIndex.Matrix.Format(aacp.aaindex.matrix)
}
\arguments{
\item{aacp.aaindex.matrix}{AAIndex AACP matrix}
}
\description{
Format AAIndex Amino Acid Pairwise Contact Potential (AACP) matrix
}
