% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContactPotentialProfiling_FragmentLibrary.R
\name{ContactPotentialProfiling_FragmentLibrary}
\alias{ContactPotentialProfiling_FragmentLibrary}
\alias{CPP_FragmentLibrary}
\title{Generate sequence fragment libraries.}
\usage{
CPP_FragmentLibrary(sequenceSet, fragLenSet = 3:8,
  maxFragDepth = 1e+05, seedSet = 1:5)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}

\item{fragLenSet}{A set of sliding window sizes. Must be between 3 and 8.}

\item{maxFragDepth}{The maximum size of the fragment library.}

\item{seedSet}{A set of random seeds.}
}
\description{
\code{CPP_FragmentLibrary} generates a formatted fragment library datatable.
}
