% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_EpitopeDataset.R
\name{DataPreparation_EpitopeDataset}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Import}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_SolveImmunogenicityContradiction}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Add_Disease}
\alias{DataPreparation_EpitopeDataset}
\alias{Epitope_Add_HLASerotypes}
\title{Compile epitope datasets retrieved from IEDB and other sources.}
\usage{
Epitope_Import(IEDBAssayFileName, OtherFileNames = NULL)

Epitope_SolveImmunogenicityContradiction(df)

Epitope_Add_Disease(df,
  IEDBEpitopeFileName = system.file("IEDB_Epitope_Healthy.csv.gz", package =
  "Repitope"))

Epitope_Add_HLASerotypes(df,
  IEDBEpitopeFileNames = system.file("IEDB_Epitope_Serotype_HLA-A01.csv.gz",
  package = "Repitope"))
}
\arguments{
\item{IEDBAssayFileName}{A set of T cell assay results obtained from IEDB.}

\item{OtherFileNames}{Other sources. Must contain a "Dataset" column indicating the data source.}

\item{df}{An epitope dataframe.}

\item{IEDBEpitopeFileName}{A set of epitopes obtained from IEDB.}

\item{IEDBEpitopeFileNames}{A set of epitope files from IEDB.}
}
\description{
Compile epitope datasets retrieved from IEDB and other sources.
}
