% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_NetMHC.R
\name{DataPreparation_NetMHC}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_PeptideExport}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_Import}
\title{Utility functions for NetMHC.}
\usage{
NetMHC_PeptideExport(peptideSet, outDir = "./NetMHC/")

NetMHC_Import(netMHCOutputFileNames, MHCType = c("MHCI", "MHCII"))
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{outDir}{A path to the output directory.}

\item{netMHCOutputFileNames}{A set of tab-delimited NetMHC result files.}

\item{MHCType}{A character string. "MHCI" or "MHCII".}
}
\description{
Utility functions for NetMHC.
}
