% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation_NetMHC.R
\name{DataPreparation_NetMHC}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_PeptideExport}
\alias{DataPreparation_NetMHC}
\alias{NetMHC_Import}
\title{Utilities for NetMHC.}
\usage{
NetMHC_PeptideExport(peptideSet, outDir = "./NetMHC/")

NetMHC_Import(netMHCFileNames, hlaSuperTypeSet = c("A01", "A02", "A03", "A24",
  "A26", "B07", "B08", "B27", "B39", "B44", "B58", "B62"))
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{outDir}{A path to the output directory.}

\item{netMHCFileNames}{A set of tab-delimited NetMHC result files.}

\item{hlaSuperTypeSet}{A set of HLA supertypes for which binding was predicted.}
}
\description{
Utilities for NetMHC.
}
