% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{EpitopeDataset}
\title{A compiled dataset of human MHC class I epitopes and ligands.}
\format{A tibble with 21146 rows and 11 variables}
\usage{
EpitopeDataset
}
\description{
A compiled dataset of human MHC class I epitopes and ligands.
}
\keyword{datasets}
