% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{MHCI_Human}
\alias{Dataset_EpitopeDataset}
\alias{MHCI_Rodents}
\alias{Dataset_EpitopeDataset}
\alias{MHCI_Primates}
\alias{Dataset_EpitopeDataset}
\alias{MHCII_Human}
\alias{Dataset_EpitopeDataset}
\alias{MHCII_Rodents}
\alias{Dataset_EpitopeDataset}
\alias{MHCII_Primates}
\title{Compiled datasets of MHC class I and class II epitopes and ligands.}
\format{data.table}
\usage{
MHCI_Human

MHCI_Rodents

MHCI_Primates

MHCII_Human

MHCII_Rodents

MHCII_Primates
}
\keyword{datasets}
