% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{EpitopeDataset}
\title{A compiled dataset of MHC class I epitopes and ligands.}
\format{A tibble with 19533 rows and 7 variables}
\usage{
EpitopeDataset
}
\description{
\code{EpitopeDataset} is a dataframe containing 6159 epitopes and 13374 MHC ligands. Entries were retrieved from various sources. The columns Cluster60 and Cluster80 contain the clustering results by the IEDB peptide clustering tool, with threshold of 60% and 80%, respectively.
}
\examples{
dplyr::filter(EpitopeDataset, Dataset=="Calis1")
dplyr::filter(EpitopeDataset, stringr::str_detect(MHC, stringr::fixed("A*02"))|stringr::str_detect(MHC, "A02")|stringr::str_detect(MHC, "A2$")|stringr::str_detect(MHC, stringr::fixed("A2|")))
}
\keyword{datasets}
