% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{EpitopeDataset}
\title{A compiled dataset of MHC class I epitopes and ligands.}
\format{A tibble with 21768 rows and 9 variables}
\usage{
EpitopeDataset
}
\description{
\code{EpitopeDataset} is a dataframe containing 7106 epitopes and 14662 MHC ligands. Entries were retrieved from various sources. The columns Cluster60 and Cluster80 contain the clustering results by the IEDB peptide clustering tool, with threshold of 60% and 80%, respectively.
}
\examples{
## Data derived from the "Calis1" dataset.
df.calis1 <- dplyr::filter(EpitopeDataset, stringr::str_detect(Dataset, "Calis1"))

## Data restricted on the HLA-A2 molecule.
df <- compressedToLongFormat(EpitopeDataset, "MHC")
df$"MHC" <- standardizeHLAString(df$"MHC")
df.hla.a2 <- df.long \%>\%
  dplyr::filter(stringr::str_detect(MHC, "A02")|stringr::str_detect(MHC, "A2$")) \%>\%
  dplyr::distinct(Peptide, .keep_all=T)
}
\keyword{datasets}
