% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_EpitopeDataset.R,
%   R/Dataset_FeatureSet.R
\docType{data}
\name{Dataset_EpitopeDataset}
\alias{Dataset_EpitopeDataset}
\alias{MHCI_Human}
\alias{MHCI_Rodents}
\alias{MHCI_Primates}
\alias{MHCII_Human}
\alias{MHCII_Rodents}
\alias{MHCII_Primates}
\alias{MHCI_Human_MinimumFeatureSet}
\alias{MHCII_Human_MinimumFeatureSet}
\title{Compiled datasets of MHC class I and class II epitopes and ligands.}
\format{data.table}
\usage{
MHCI_Human

MHCI_Rodents

MHCI_Primates

MHCII_Human

MHCII_Rodents

MHCII_Primates

MHCI_Human_MinimumFeatureSet

MHCII_Human_MinimumFeatureSet
}
\description{
Datasets were compiled from source files attached in the package using \code{Epitope_Import}.
}
\keyword{datasets}
