% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_TCRDatasets.R
\docType{data}
\name{Dataset_TCRDatasets}
\alias{Dataset_TCRDatasets}
\alias{TCRSet}
\title{Compiled datasets of human TCR beta CDR3 sequences}
\format{A character vector.}
\usage{
TCRSet
}
\description{
\code{TCRSet} is a combined human TCR repertoire derived from unsorted T cells. TCR clonotypes observed in at least 22 out of 218 datasets were retained.\cr
\code{TCRSet_CD8} is a combined human CD8+ TCR repertoire. TCR clonotypes observed in at least 2 out of 6 donors were retained.\cr
\code{TCRSet_PubClones10000} is a public TCR repertoire returned from \code{publicClonotypeAnalysis}. The size threshold was set 10000.\cr
\code{TCRSet_CD8_PubClones10000} is a public CD8+ TCR repertoire returned from \code{publicClonotypeAnalysis}. The size threshold was set 10000.\cr
For details, type \code{help("TCRAnalysis_PublicClonotypes")}.
}
\examples{
## Calculate features.
FeatureDFList_CD8 <- Features(
  peptideSet=EpitopeDataset$"Peptide",
  TCRSet=TCRSet_CD8_PubClones10000,
  aaIndexIDSet="all",
  alignTypeSet="global-local",
  fragLenSet=3:8,
  TCRFragDepthSet=10000,
  seedSet=1:5
)

}
\keyword{datasets}
