% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_TCRDatasets.R
\docType{data}
\name{Dataset_TCRDatasets}
\alias{Dataset_TCRDatasets}
\alias{TCRSet_Public}
\title{A compiled dataset of human TCR beta CDR3 sequences}
\format{A character vector containing 191,326 sequences.}
\usage{
TCRSet_Public
}
\description{
\code{TCRSet_Public} is a set of TCR beta CDR3 clonotype sequences commonly observed in multiple TCR repertoire datasets obtianed from healthy human individuals.
}
\keyword{datasets}
