% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpitopePrioritization.R
\name{EpitopePrioritization}
\alias{EpitopePrioritization}
\title{Epitope prioritization based on immunogenicity scores and escape potentials.}
\usage{
EpitopePrioritization(featureDF, metadataDF, peptideSet, fragLib,
  aaIndexIDSet = "all", fragLenSet = 3:8, fragDepth = 10000,
  fragLibType = "Weighted", featureSet = NULL, seedSet = 1:5,
  coreN = parallel::detectCores(logical = F),
  outDir = paste0("./Output_", format(Sys.time(),
  "\%Y.\%m.\%d.\%H.\%M.\%S")))
}
\arguments{
\item{featureDF}{A pre-computed feature dataframe for model construction.}

\item{metadataDF}{A metadata dataframe, consisting only of "Peptide" and "Immunogenicity" columns, for model construction.}

\item{peptideSet}{A set of new peptide sequences.}

\item{fragLib}{Either the dataframe of fragment libraries generated by \code{CPP_FragmentLibrary} or the path to the fst file.}

\item{aaIndexIDSet}{A set of AAIndex IDs indicating the AACP scales to be used. Set "all" to shortcut the selection of all available AACP scales.}

\item{fragLenSet}{A set of sliding window sizes.}

\item{fragDepth}{A fragment library size. Must be an integer vector of length one.}

\item{fragLibType}{A string indicating the types of fragment libraries to be used. Must be a character vector of length one.}

\item{featureSet}{A minimum set of features. Combinations of fragment lengths and AAIndex IDs are internally extracted for calculating CPPs.}

\item{seedSet}{A set of random seeds.}

\item{coreN}{The number of cores to be used for parallelization.}

\item{outDir}{Destination directory to save the results.}
}
\description{
Epitope prioritization based on immunogenicity scores and escape potentials.
}
