% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{Features}
\alias{Features}
\title{Generate peptide features for immunogenicity prediction.}
\usage{
Features(peptideSet, TCRSet, fragLenSet = 3:8, aaIndexIDSet = "all",
  alignTypeSet = "global-local", TCRFragDepthSet = 10000, seedSet = 1:5,
  coreN = parallel::detectCores())
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a list, it must be the same length with the seedSet.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched. This should be kept constant for comparison.}

\item{seedSet}{A set of random seeds.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
\code{Features_PeptideDescriptor} calculates descriptive statistics using peptide descriptors.\cr
\code{Features_rTPCP} calculates descriptive statistics of repertoire-wide TCR-peptide pairwise contact potentials.\cr
\code{Features} is a wrapper function, generating a list of dataframes of peptide features calculated with different parameter sets.
}
