% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{Features_PeptDesc}
\alias{Features_PeptDesc}
\alias{Features_CPP}
\alias{Features_CPP}
\alias{Features}
\title{Generate peptide features.}
\usage{
Features_PeptDesc(peptideSet, fragLenSet = 3:8)

Features_CPP(peptideSet, aaIndexIDSet = "all", fragLenSet = 3:8,
  fragDepthSet = 10000, seedSet = 1:5, coreN = parallel::detectCores(),
  tmpDir = tempdir())

Features(peptideSet, aaIndexIDSet = "all", fragLenSet = 3:8,
  fragDepthSet = 10000, seedSet = 1:5, coreN = parallel::detectCores(),
  tmpDir = tempdir())
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{fragLenSet}{A set of sliding window sizes. Must be between 3 and 8.}

\item{aaIndexIDSet}{A set of AAIndex IDs indicating the AACP scales to be used. Set "all" to shortcut the selection of all available AACP scales.}

\item{seedSet}{A set of random seeds.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}

\item{tmpDir}{Destination directory to save intermediate files.}
}
\description{
\code{Features_PeptDesc} calculates descriptive statistics of peptide descriptors.\cr
\code{Features_CPP} calculates descriptive statistics of interresidue pairwise contact potential profiles.\cr
\code{Features} is a wrapper function.
}
