% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R, R/Features_ErrorCorrect.R,
%   R/Features_PeptideDescriptor.R, R/Features_rTPCP.R
\name{Features}
\alias{Features}
\alias{Features}
\alias{Features_ErrorCorrect}
\alias{Features}
\alias{Features_PeptideDescriptor}
\alias{Features}
\alias{Features_rTPCP}
\title{Generate peptide features for immunogenicity prediction.}
\usage{
Features(peptideSet, TCRSet, fragLenSet = 3:8, aaIndexIDSet = "all",
  alignTypeSet = "global-local", TCRFragDepthSet = 10000, seedSet = 1:5)

Features_ErrorCorrect(featureDFList, TCRSet, fragLenSet = 3:8,
  aaIndexIDSet = "all", alignTypeSet = "global-local",
  TCRFragDepthSet = 10000, seedSet = 1:5)

Features_PeptideDescriptor(peptideSet, fragLenSet = 3:8)

Features_rTPCP(peptideSet, TCRSet, fragLenSet = 3:8, aaIndexIDSet = "all",
  alignTypeSet = "global-local", TCRFragDepthSet = 10000, seedSet = 1:5)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a list, it must be the same length with the seedSet.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched. This should be kept constant for comparison.}

\item{seedSet}{A set of random seeds.}

\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a list, it must be the same length with the seedSet.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched. This should be kept constant for comparison.}

\item{seedSet}{A set of random seeds.}

\item{peptideSet}{A set of peptide sequences.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{peptideSet}{A set of peptide sequences.}

\item{TCRSet}{Either a set of TCR sequences (as a character vector) or a list of sets of TCR sequences. If provided as a character vector, TCR sequences are randomly down-sampled to 10000, using \code{seedSet} to create a list. If provided as a list, it must be the same length with the seedSet.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched. This should be kept constant for comparison.}

\item{seedSet}{A set of random seeds.}
}
\description{
A wrapper of \code{Features_PeptideDescriptor} and \code{Features_rTPCP}, generating a list of dataframes of peptide features calculated with different parameter sets.

Correct errors in the peptide feature dataframes.

Calculates descriptive statistics using peptide descriptors.

Calculates descriptive statistics of repertoire-wide TCR-peptide pairwise contact potentials.
}
