% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Importance.R
\name{Features_Importance}
\alias{Features_Importance}
\title{Feature selection based on their importances.}
\usage{
Features_Importance(preprocessedDFList, featureN = 100,
  coreN = parallel::detectCores())
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess} or \code{Features_CorFilter}.}

\item{featureN}{The number of the features that should be retained. Can be \code{NULL} to disable feature selection.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
Feature selection based on their importances.
}
