% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Importance.R
\name{Features_Importance}
\alias{Features_Importance}
\alias{Features_Importance}
\alias{Features_Importance_Reduce}
\title{Feature selection based on their importances.}
\usage{
Features_Importance(preprocessedDFList, featureN = 100)

Features_Importance_Reduce(featureImportanceDFList, featureN = 30)
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess} or \code{Features_CorFilter}.}

\item{featureN}{The number of the features that should be retained. Can be \code{NULL} to disable feature selection.}

\item{featureImportanceDFList}{An output generated by \code{Features_Importance}.}
}
\description{
\code{Features_Importance} calculates feature importances using Random Forest algorithm.\cr
\code{Features_Importance_Reduce} reduces the number of the retained features.\cr
}
