% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_MinimumFeatures.R
\name{Features_MinimumFeatures}
\alias{Features_MinimumFeatures}
\title{The minimum set of the most important features.}
\usage{
Features_MinimumFeatures(featureDFList, metadataDF, seedSet = 1:5,
  corThreshold = 0.75, featureNSet = 100, criteria = "intersect",
  returnImpDFList = T)
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}, or imported by \code{readFeatureDFList}.}

\item{metadataDF}{A dataframe containing metadata, consisting of "Peptide", "Immunogenicity", and "Cluster" columns.}

\item{seedSet}{A set of random seeds.}

\item{corThreshold}{The threshold of correlation to eliminate features.}

\item{featureNSet}{The number of the features that should be retained. Can also be a vector of numbers. In that case, a list of minimum sets of features would be returned.}

\item{criteria}{The criteria of feature extraction. Can either be "intersect" or "union".}

\item{returnImpDFList}{Logical. Whether the lists of feature importance dataframes should be returned as well.}
}
\description{
\code{Features_MinimumFeatures} automates the feature selection workflows, i.e. preprocessing, filtering by correlation, filtering by importance, extract shared features from ascending vs descending models. \cr
}
