% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_PeptideDescriptor.R
\name{Features_PeptideDescriptor}
\alias{Features_PeptideDescriptor}
\title{Descriptive statistics of peptide fragment descriptors.}
\usage{
Features_PeptideDescriptor(peptideSet, fragLenSet = 5)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}
}
\description{
Calculates descriptive statistics using peptide descriptors.
}
