% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocessing.R
\name{Features_Preprocessing}
\alias{Features_Preprocessing}
\alias{Features_Preprocess}
\alias{Features_Preprocessing}
\alias{Features_CorFilter}
\alias{Features_Preprocessing}
\alias{Features_FeatureSelection}
\title{Preprocess peptide feature dataframes for machine learning.}
\usage{
Features_Preprocess(featureDFList, metadataDFList)

Features_CorFilter(preprocessedDTList, corThreshold = 0.75,
  coreN = parallel::detectCores())

Features_FeatureSelection(corfilteredDTList, coreN = parallel::detectCores())
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{metadataDFList}{A list of metadata dataframes labeled as "Peptide", "Immunogenicity", and "Cluster". No extra metadata is allowed. Feature dataframes are downsized based on the "Peptide" column in this metadata. In other words, peptides not contained in the metadata dataframe are omitted.}

\item{preprocessedDTList}{A list of feature dataframes generated by \code{Features_Preprocess}.}

\item{corThreshold}{The threshold for correlation-based feature elimination.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}

\item{corfilterDTList}{A list of feature dataframes generated by \code{Features_CorFilter}.}
}
\description{
Preprocess peptide feature dataframes for machine learning.
}
