% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocessing.R
\name{Features_Preprocessing}
\alias{Features_Preprocessing}
\alias{Features_Preprocess}
\alias{Features_Preprocessing}
\alias{Features_CorFilter}
\alias{Features_Preprocessing}
\alias{Features_FeatureSelect}
\title{Preprocess peptide feature dataframes for machine learning.}
\usage{
Features_Preprocess(featureDFList, metadataDFList)

Features_CorFilter(preprocessedDFList, corThreshold = 0.75,
  coreN = parallel::detectCores())

Features_FeatureSelect(preprocessedDFList, featureN = 100,
  coreN = parallel::detectCores())
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{metadataDFList}{A list of metadata dataframes labeled as "Peptide", "Immunogenicity", and "Cluster". No extra metadata is allowed. Feature dataframes are downsized based on the "Peptide" column in this metadata. In other words, peptides not contained in the metadata dataframe are omitted.}

\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess} or \code{Features_CorFilter}.}

\item{corThreshold}{The threshold for correlation-based feature elimination.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}

\item{featureN}{The number of the features that should be retained.}
}
\description{
Preprocess peptide feature dataframes for machine learning.
}
