% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocess.R
\name{Features_Preprocessing}
\alias{Features_Preprocessing}
\alias{Features_Preprocess_ByReference}
\alias{Features_Preprocessing}
\alias{Features_SplitDF}
\alias{Features_Preprocessing}
\alias{Features_Preprocess}
\alias{Features_Preprocessing}
\alias{Features_FeatureSelection}
\title{Preprocess peptide feature dataframes for machine learning.}
\usage{
Features_Preprocess_ByReference(featureDFList, metadataDFList,
  preprocessedDFList_reference)

Features_SplitDF(featureDFList, metadataDFList)

Features_Preprocess(splitDFList)

Features_FeatureSelection(preprocessedDFList, coreN = parallel::detectCores())
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}.}

\item{metadataDFList}{A list of metadata dataframes labeled as "Peptide", "Immunogenicity", and "Cluster". No extra metadata is allowed. Feature dataframes are downsized based on the "Peptide" column in this metadata. In other words, peptides not contained in the metadata dataframe are omitted.}

\item{preprocessedDFList_reference}{A list of feature dataframes generated by \code{Features_Preprocess} or \code{Features_RFE} for reference.}

\item{splitDFList}{A list of feature dataframes generated by \code{Features_SplitDF}.}

\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess}.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
Preprocess peptide feature dataframes for machine learning.
}
