% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features_Preprocessing.R
\name{Features_Preprocessing}
\alias{Features_Preprocessing}
\alias{Features_Preprocess}
\alias{Features_CorFilter}
\alias{Features_FeatureSelect}
\alias{inverseColOrderDFList}
\title{Feature preprocessing.}
\usage{
Features_Preprocess(featureDFList, metadataDF, seedSet = 1:5)

Features_CorFilter(preprocessedDFList, corThreshold = 0.75,
  coreN = parallel::detectCores(logical = F))

Features_FeatureSelect(preprocessedDFList, featureN = 100,
  coreN = parallel::detectCores(logical = F))

inverseColOrderDFList(preprocessedDFList)
}
\arguments{
\item{featureDFList}{A list of feature dataframes generated by \code{Features}, or imported by \code{readFeatureDFList}.}

\item{metadataDF}{A dataframe containing metadata, consisting of "Peptide", "Immunogenicity", and "Cluster" columns.}

\item{seedSet}{A set of random seeds.}

\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess} or \code{Features_CorFilter}.}

\item{corThreshold}{The threshold of correlation to eliminate features.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to disable.}

\item{featureN}{The number of features to be retained.}
}
\description{
\code{Features_Preprocess} does preprocessing, i.e. centering and rescaling using the internally separated training subdata.\cr
\code{Features_CorFilter} filters highly correlated features. Internally using \code{parCor} to do parallelized calculation of a correlational matrix.\cr
\code{Features_FeatureSelect} selects important features. See also \code{Features_Importance}.\cr
\code{inverseColOrderDFList} takes \code{preprocessedDFList} and reverses the column order of the datatable.
}
