% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity.R
\name{Immunogenicity}
\alias{Immunogenicity}
\alias{Immunogenicity_Benchmark}
\title{Immunogenicity prediction.}
\usage{
Immunogenicity(preprocessedDFList, featureSet = "all",
  destDir = "./Immunogenicity/", outputHeader = "Output",
  maxJavaMemory = "6G", coreN = parallel::detectCores())

Immunogenicity_Benchmark(preprocessedDFList, learnerSet = NULL,
  omitLearnerSet = NULL, destDir = "./Benchmark/", maxJavaMemory = "6G",
  coreN = parallel::detectCores())
}
\arguments{
\item{preprocessedDFList}{An output generated by \code{Features_Preprocess}, \code{Features_CorFilter}, or \code{Features_FeatureSelect}.}

\item{featureSet}{A set of features to be used for model training. If "all", all features in the training dataframe except "DataType", "Immunogenicity", "Peptide", and "Cluster" will be used.}

\item{destDir}{A directory for loading & exporting data.}

\item{outputHeader}{A file/folder name header.}

\item{maxJavaMemory}{The upper limit of memory for Java virtual machine.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}

\item{learnerSet}{(Optional) A set of learner strings to be used for model benchmarking. Set \code{NULL} to ignore.}

\item{omitLearnerSet}{(Optional) A set of learner strings to be omitted from model benchmarking. Set \code{NULL} to ignore.}
}
\description{
\code{Immunogenicity} is a wrapper function for training and evaluating an extraTrees classifier.\cr
\code{Immunogenicity_Benchmark} is a wrapper function for benchmarking classifiers implemented in the \code{mlr} package.\cr
}
