% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immunogenicity_Score.R
\name{Immunogenicity_Score}
\alias{Immunogenicity_Score}
\title{Immunogenicity score.}
\usage{
Immunogenicity_Score(featureDFFileNames_train,
  featureDFFileNames_predict = NULL, metadata_train, featureSet = "all",
  seedSet = 1:10, maxJavaMemory = "6G", coreN = parallel::detectCores())
}
\arguments{
\item{featureDFFileNames_train}{A character vector of the fst file names of feature dataframes for model training.}

\item{featureDFFileNames_predict}{A character vector of the fst file names of feature dataframes for predicting immunogenicity scores. If \code{NULL}, model training data was splitted into training and prediction subdatasets.}

\item{metadata_train}{A dataframe containig "Peptide" and "Immunogenicity" columns for model training.}

\item{featureSet}{A set of features for model training.}

\item{seedSet}{A set of random seeds for bootstrapping.}

\item{maxJavaMemory}{The upper limit of memory for Java virtual machine in megabytes.}

\item{coreN}{The number of cores to be used for parallelization. Set \code{NULL} to skip parallelization.}
}
\description{
\code{Immunogenicity_Score} calculates immunogenicity scores.
}
