% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmunogenicityPrediction.R
\name{ImmunogenicityPrediction}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity_AutoML}
\alias{ImmunogenicityPrediction}
\alias{Immunogenicity_Prediction}
\title{Immunogenicity prediction using H2O AutoML instances}
\usage{
Immunogenicity(preprocessedDFList, destDir = "./Results/",
  max_mem_size = "6G", nthreads = 6)

Immunogenicity_AutoML(trainDF, testDF, validDF, destDir = "./Results/",
  LeaderBoardName = "LeaderBoard.csv", H2OModelName = "BestH2OModel",
  seed = 12345, max_mem_size = "6G", nthreads = 6)

Immunogenicity_Prediction(evalDF, destDir = "./Results/",
  H2OModelName = "BestH2OModel", PredictionHeader = "Predictions_",
  seed = 12345, max_mem_size = "6G", nthreads = 6)
}
\arguments{
\item{preprocessedDFList}{A list of feature dataframes generated by \code{Features_Preprocess}.}

\item{destDir}{A directory for saving outputs.}

\item{max_mem_size}{The upper limit of memory for H2O Java machine learning engine.}

\item{nthreads}{The number of threads for H2O Java machine learning engine.}

\item{trainDF}{A dataframe for model training.}

\item{testDF}{A dataframe for hold-out validation.}

\item{validDF}{A dataframe for leaderboard scoring.}

\item{LeaderBoardName}{The file name of the leader board to be saved. The file format must be CSV. If you want to skip saving, set \code{NULL}.}

\item{H2OModelName}{The file name of the best H2O model to be saved. No file extension is required. If you want to skip saving, set \code{NULL}.}

\item{seed}{A random seed.}

\item{evalDF}{A dataframe for evaluation.}

\item{PredictionHeader}{The file name header of the prediction results to be saved.}
}
\description{
\code{Immunogenicity} is a wrapper function.
}
