% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicityPrediction_Features.R
\name{immunogenicityPrediction_Features}
\alias{immunogenicityPrediction_Features}
\title{Generate peptide features for immunogenicity prediction.}
\usage{
immunogenicityPrediction_Features(peptideSet, TCRSet, aaIndexIDSet = "all",
  alignTypeSet = "global-local", fragLenSet = 5, TCRFragDepthSet = 10000,
  seedSet = 1:5)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{TCRSet}{A set of TCR sequences.}

\item{aaIndexIDSet}{A set of AAIndexIDs indicating the pairwise matching score matrix to be used. A set of AAIndex-derived matrices can be retrieved by \code{AACPMatrix}. Set "all" to select all AAIndexIDs.}

\item{alignTypeSet}{A set of alignment-type strings directly passed to the \code{type} argument of the \code{pairwiseAlignment} function in the \code{Biostrings} package.}

\item{fragLenSet}{A set of the lengths of TCR sequence fragments to be matched against the peptide set.}

\item{TCRFragDepthSet}{A set of the numbers of TCR fragments to be matched. This should be kept constant for comparison.}

\item{seedSet}{A set of random seeds.}
}
\description{
A wrapper function of \code{peptideDescriptorAnalysis} and \code{rTPCPAnalysis}.
}
