% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicityPrediction_MachineLearning.R
\name{immunogenicityPrediction_MachineLearning}
\alias{immunogenicityPrediction_MachineLearning}
\alias{immunogenicityPrediction_MachineLearning.DataFormat}
\alias{immunogenicityPrediction_MachineLearning}
\alias{immunogenicityPrediction_MachineLearning.Single}
\alias{immunogenicityPrediction_MachineLearning}
\title{Machine learning for immunogenicity prediction}
\usage{
immunogenicityPrediction_MachineLearning.DataFormat(featureDF, seed = 12345)

immunogenicityPrediction_MachineLearning.Single(formattedDataList,
  trainControlOptions, mlAlgorithm, mlAlgorithmLabel, seed = 12345,
  outDir = "./classifiers/")

immunogenicityPrediction_MachineLearning(featureDFSet, metadataDF,
  trainControlOptions, mlAlgorithmSet, mlAlgorithmLabelSet, seedSet = 1:5,
  outDir = "./classifiers/")
}
\arguments{
\item{featureDF}{A feature dataframe. A "Peptide" column is required.}

\item{seed}{A random seed.}

\item{formattedDataList}{A formatted list of dataframes returned by \code{immunogenicityPrediction_MachineLearning.DataFormat}.}

\item{trainControlOptions}{A trainControl object.}

\item{mlAlgorithm}{A string indicating the caret machine learning algorithm.}

\item{mlAlgorithmLabel}{A label for the caret machine learning algorithm.}

\item{outDir}{A directory for exporting the results.}

\item{featureDFSet}{A set of feature dataframes.}

\item{metadataDF}{A metadata dataframe labeled as "Peptide", "Immunogenicity", "MHC", and "Cluster". No extra metadata is allowed (mistaken as features).}

\item{mlAlgorithmSet}{A set of strings indicating the caret machine learning algorithms.}

\item{mlAlgorithmLabelSet}{A set of labels for the caret machine learning algorithms.}

\item{seedSet}{A set of random seeds.}
}
\description{
\code{immunogenicityPrediction_MachineLearning.DataFormat} generates machine-learning-ready datasets.\cr
\code{immunogenicityPrediction_MachineLearning.Single} performs machine learning with a single algorithm and a single random seed.\cr
\code{immunogenicityPrediction_MachineLearning} performs machine learning with multiple combinations of algorithms and random seeds.
}
