% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsillicoMutagenesis.R
\name{InsillicoMutagenesis}
\alias{InsillicoMutagenesis}
\alias{InSillicoMutagenesis}
\alias{InsillicoMutagenesis}
\alias{InsillicoMutagenesis_GenerateIntermediates}
\alias{InsillicoMutagenesis}
\alias{InsillicoMutagenesis_PairwiseSimilarityNetwork}
\title{In sillico mutagenesis analysis.}
\usage{
InSillicoMutagenesis(peptideSet, union = T, coreN = NULL)

InsillicoMutagenesis_GenerateIntermediates(peptidePair)

InsillicoMutagenesis_PairwiseSimilarityNetwork(peptidePairDF, coreN = NULL)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{union}{Logical. Whether the output list should be combined.}

\item{coreN}{The number of threads.}

\item{peptidePair}{A pair of peptide sequences.}

\item{peptidePairDF}{A dataframe which has two columns of paired peptide sequences.}
}
\description{
\code{InSillicoMutagenesis} generates single-mutated peptide sequences.
\code{InsillicoMutagenesis_GenerateIntermediates} generates mutational intermediates between two peptide sequences.
\code{InsillicoMutagenesis_PairwiseSimilarityNetwork} generates single-mutated peptide sequences and construct similarity networks.
}
