% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsillicoMutagenesis.R
\name{InsillicoMutagenesis}
\alias{InsillicoMutagenesis}
\alias{InsillicoMutagenesis_GenerateIntermediates}
\title{In sillico mutagenesis analysis.}
\usage{
InsillicoMutagenesis_GenerateIntermediates(peptide.start, peptide.goal)
}
\arguments{
\item{peptide.start}{A peptide sequence.}

\item{peptide.goal}{A peptide sequence.}
}
\description{
In sillico mutagenesis analysis.
}
