% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InSillicoMutagenesis.R
\name{InSillicoMutagenesis}
\alias{InSillicoMutagenesis}
\alias{InSillicoMutagenesis}
\alias{InSillicoMutagenesis_GenerateIntermediates}
\title{In sillico mutagenesis analysis.}
\usage{
InSillicoMutagenesis(peptideSet, coreN = parallel::detectCores())

InSillicoMutagenesis_GenerateIntermediates(peptidePair,
  coreN = parallel::detectCores())
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{coreN}{The number of threads for parallelization.}

\item{peptidePair}{A pair of peptide sequences.}
}
\description{
\code{InSillicoMutagenesis} generates single-mutated peptide sequences.
\code{InSillicoMutagenesis_GenerateIntermediates} generates mutational intermediates between two peptide sequences.
}
