% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsillicoMutagenesis.R
\name{InsillicoMutagenesis}
\alias{InsillicoMutagenesis}
\alias{InSillicoMutagenesis}
\title{In sillico mutagenesis analysis.}
\usage{
InSillicoMutagenesis(peptideSet)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{peptide.start}{A peptide sequence.}

\item{peptide.goal}{A peptide sequence.}
}
\description{
In sillico mutagenesis analysis.
}
