% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrTPCP.R
\name{mrTPCP.fixedWindow}
\alias{mrTPCP.fixedWindow}
\title{Modified repertoire-wide TCR-peptide contact profile (mrTPCP) analysis with a fixed window size}
\usage{
mrTPCP.fixedWindow(peptList, rept = NULL, wind = 4,
  aa.index.id = "MIYS990106")
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences.}

\item{rept}{A character vector containing input TCR CDR3 sequences. A default is NULL; in this case, a reference CD8 TCR repertoire used in the original paper will be used. Alternatively, users can provide a character vector containing a CDR3 sequence repertoire of interest.}

\item{wind}{A size of the window.}

\item{aa.index.id}{The AAIndex name of the AACP scale to be used. Currently, only one AAIndex name is supported.}
}
\description{
Modified repertoire-wide TCR-peptide contact profile (mrTPCP) analysis with a fixed window size
}
