% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrTPCP.R
\name{mrTPCP.fixedWindow}
\alias{mrTPCP.fixedWindow}
\title{Calculate modified repertoire-wide TCR-peptide contact profile (mrTPCP) variables with a fixed window size}
\usage{
mrTPCP.fixedWindow(peptList, rept, wind, aa.index.id = c("MIYS990106"))
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences}

\item{rept}{A character vector containing input TCR CDR3 sequences}

\item{wind}{A size of window}

\item{aa.index.id}{A character vector of AAIndex names to be used}
}
\description{
Calculate modified repertoire-wide TCR-peptide contact profile (mrTPCP) variables with a fixed window size
}
