% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrTPCP.R
\name{mrTPCP}
\alias{mrTPCP}
\title{Modified repertoire-wide TCR-peptide contact profile (mrTPCP) analysis}
\usage{
mrTPCP(peptList, repertoire = NULL, winds = 4:5,
  aa.index.id = "MIYS990106")
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences. Peptides must be of the same length.}

\item{repertoire}{A character vector containing input TCR CDR3 sequences. A default is NULL; in this case, a reference CD8 TCR repertoire used in the original paper will be used. Alternatively, users can provide a character vector containing a CDR3 sequence repertoire of interest.}

\item{winds}{A vector of the sizes of the window.}

\item{aa.index.id}{The AAIndex name of the AACP scale to be used. Currently, only one AAIndex name is supported.}
}
\description{
\code{mrTPCP} returns the modified repertoire-wide TCR-peptide contact profile (mrTPCP) variables.
}
