% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrTPCP.R
\name{mrTPCP}
\alias{mrTPCP}
\title{Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables}
\usage{
mrTPCP(peptList, repertoire, winds = 4:5, aa.index.id = c("MIYS990106"))
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences}

\item{repertoire}{A character vector containing input TCR CDR3 sequences}

\item{winds}{A vector of the sizes of window}

\item{aa.index.id}{A character vector of AAIndex names to be used}
}
\description{
Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables
}
