% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborNetwork.R
\name{NeighborNetwork}
\alias{NeighborNetwork}
\alias{neighborNetwork}
\title{Neighbor network analysis.}
\usage{
neighborNetwork(peptideSet, numSet = NULL, directed = F,
  weighted = F, coreN = parallel::detectCores(logical = F))
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{numSet}{An attribute for the vertices.}

\item{directed}{Should the network be converted from undirected to directed? Directions are determined using the \code{numSet} provided.}

\item{weighted}{Should the network be converted to weihted? Edge weights are determined using the \code{numSet} provided.}

\item{coreN}{The number of cores to be used for parallelization.}
}
\description{
Neighbor network analysis.
}
