% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunogenicityPrediction_PeptideDescriptor.R
\name{peptideDescriptor}
\alias{peptideDescriptor}
\alias{peptideDescriptorAnalysis}
\title{Descriptive statistics of peptide fragment descriptors.}
\usage{
peptideDescriptorAnalysis(peptideSet, fragLen = 5)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{fragLen}{The length of TCR sequence fragments to be matched against the peptide set.}
}
\description{
Descriptive statistics of peptide fragment descriptors.
}
