% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTPCP.R
\name{rTPCP.fixedWindow}
\alias{rTPCP.fixedWindow}
\title{Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables with a fixed window size}
\usage{
rTPCP.fixedWindow(peptList, rept, wind, aa.index.id = c("MIYS990106"))
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences}

\item{rept}{A character vector containing input TCR CDR3 sequences}

\item{wind}{A size of window}

\item{aa.index.id}{A character vector of AAIndex names to be used}
}
\description{
Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables with a fixed window size
}
