% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTPCP.R
\name{rTPCP}
\alias{rTPCP}
\title{Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables}
\usage{
rTPCP(peptList, repertoire = NULL, winds = 4:5,
  aa.index.id = c("MIYS990106"))
}
\arguments{
\item{peptList}{A character vector containing input peptide sequences}

\item{repertoire}{A character vector containing input TCR CDR3 sequences}

\item{winds}{A vector of the sizes of window}

\item{aa.index.id}{A character vector of AAIndex names to be used. Can also be NULL or "all", and all AAIndex scales will be used.}
}
\description{
Calculate repertoire-wide TCR-peptide contact profile (rTPCP) variables
}
