% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveCurrentGraphicPDF}
\alias{saveCurrentGraphicPDF}
\title{Plot saving utilities}
\usage{
saveCurrentGraphicPDF(outputFileName, width, height, pointsize = 12,
  family = "Helvetica", useDingbats = F)
}
\arguments{
\item{outputFileName}{The name of the exported PDF file}

\item{width}{A plot width}

\item{height}{A plot height}

\item{pointsize}{A pointsize}

\item{family}{A font family}

\item{useDingbats}{Logical indicating whether points shuold be replaced with the Dingbat font.}
}
\description{
\code{saveCurrentGraphicPDF} saves the plot currently on the graphics device to a PDF file.
}
