% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCRAnalysis_PublicClonotypes.R
\name{TCRAnalysis_PublicClonotypes}
\alias{TCRAnalysis_PublicClonotypes}
\alias{publicClonotypeAnalysis}
\title{Public XCR clonotypes in the similarity network}
\usage{
publicClonotypeAnalysis(simNet, sizeThresholdSet = 10000,
  outputFileHeader = "./Graph/Graph_")
}
\arguments{
\item{simNet}{A similarity network object.}

\item{sizeThresholdSet}{A set of integers specifying the maximum numbers of clonotypes retained as public clonotypes. Based on the degree distribution, appropriate degree threshold will be internally calculated.}

\item{outputFileHeader}{A path for saving the outputs from the public clonotype analysis.}
}
\description{
Public XCR clonotypes in the similarity network
}
