% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Utility}
\alias{Utility}
\alias{sequenceFilter}
\alias{Utility}
\alias{sequenceSlidingWindow}
\alias{Utility}
\alias{compressedToLongFormat}
\title{Miscellaneous utility functions.}
\usage{
sequenceFilter(sequenceSet)

sequenceSlidingWindow(sequenceSet, windowSize)

compressedToLongFormat(df, compressedColumnName)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}

\item{windowSize}{A size of the sliding window.}

\item{df}{A dataframe like \code{Repitope::EpitopeDataset} which has compressed columns.}

\item{compressedColumnName}{A string indicating the names of the compressed column to be converted into a long format.}
}
\description{
\code{sequenceFilter} filters amino acid sequences so that those containing non-standard letters are excluded.\cr
\code{sequenceSlidingWindow} does a sliding window with a fixed window size.\cr
\code{compressedToLongFormat} converts a compresed dataframe into a long-format dataframe. The compressed strings should be separated by "|".\cr
}
