% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParCor.R
\name{Utility_ParCor}
\alias{Utility_ParCor}
\alias{parCor}
\title{Parallelized correlation matrix calculation.}
\usage{
parCor(mat, num_splits = parallel::detectCores(), verbose = 2)
}
\arguments{
\item{mat}{A matrix. A data.frame or a data.table is internally converted into a matrix.}

\item{num_splits}{The number of chunk blocks. Default setting is the number of core (for parallelization).}

\item{verbose}{Either 0, 1, 2, or TRUE/FALSE.}
}
\description{
Parallelized correlation matrix calculation.
}
