% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParCor.R
\name{Utility_ParCor}
\alias{Utility_ParCor}
\alias{parCor}
\title{Parallelized correlation matrix calculation.}
\usage{
parCor(mat, coreN = parallel::detectCores(), verbose = T)
}
\arguments{
\item{mat}{A matrix. A data.frame or a data.table is internally converted into a matrix.}

\item{coreN}{The number of threads for parallelization. Disable by setting \code{NULL}.}

\item{verbose}{Logical. Whether the progress should be printed in the colsole?}
}
\description{
Parallelized correlation matrix calculation.
}
