% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_ParCor.R
\name{Utility_ParCor}
\alias{Utility_ParCor}
\alias{parCor}
\title{Parallelized correlation matrix computation.}
\usage{
parCor(mat, nblocks = 10, coreN = parallel:::detectCores(logical = F))
}
\arguments{
\item{mat}{A matrix, a data.frame, or a data.table.}

\item{nblocks}{The number of sub-matrices.}

\item{coreN}{The number of threads for parallelization. Disable by setting \code{NULL}.}
}
\description{
Parallelized correlation matrix computation.
}
