% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Sequence.R
\name{Utility_Sequence}
\alias{Utility_Sequence}
\alias{sequenceFilter}
\alias{sequenceSlidingWindow}
\alias{InSilicoMutagenesis}
\title{Utility functions for sequence analysis.}
\usage{
sequenceFilter(sequenceSet)

sequenceSlidingWindow(sequenceSet, windowSize)

InSilicoMutagenesis(sequenceSet)
}
\arguments{
\item{sequenceSet}{A set of amino acid sequences.}

\item{windowSize}{A size of the sliding window.}
}
\description{
\code{sequenceFilter} filters amino acid sequences so that those containing non-standard letters are excluded.\cr
\code{sequenceSlidingWindow} does a sliding window with a fixed window size.\cr
\code{InSilicoMutagenesis} generates single-aa-substituted sequences. Currently insertions and deletions are not supported.\cr
}
