% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_VennDiagram.R
\name{Utility_VennDiagram}
\alias{Utility_VennDiagram}
\alias{vennDiagram}
\title{Venn diagram.}
\usage{
vennDiagram(x, colors = (ggsci::pal_d3())(length(x)),
  show_category_names = T)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding to a separate circle in the Venn diagram. Currently, only a list of length one to five is supported.}

\item{colors}{A set of colors for the venn diagram.}

\item{show_category_names}{Logical. Whether the names of \code{x} should be printed?}
}
\description{
Venn diagram.
}
